🚚 refactor: move image files to chat
