+++ src/store/file/selectors.ts
@@ -1,4 +1,4 @@
-import { filesSelectors as imageFilesSelectors } from './slices/images';
+import { filesSelectors as imageFilesSelectors } from './slices/chat';
 import { ttsFilesSelectors } from './slices/tts';
 
 export const filesSelectors = {
